[💫 Improvements] Import more cameras tags, improve error messages
Minor improvements to handle somes tags cases better (for direction and
height)
When there is multiple directions (usually means multiple cameras), we
keep the first one and ignore the others. It is a behavior that should
be improved later but allow to display at least one field of view
instead of none.