[✨ Features] Add quatile computation and auto-assignement. Improve performances
Using the QuadTiles system, we assign every building and camera to a
tile. Suppression of the old system absed on SQL functions to use
instead a pre-generated quadtile grid imported from a SQL dump when
creating the database structure.

Get rid of the object_count attribute on tiles that was useless.

Use locators in osm2pgsql to assing buildings to a tile.

Use locators to find buildings associated to cameras and only import
them in the database (lighter database and better perormances). To do so
we need to import the cameras first, change made in the install.sh
procedure.