[⚡ Performances] Compute clusters backend side
Instead of sending all camera location and create clusters into the
frontend, we now create clusters into the backend for performance
issues. Depending if the data is a cluster or a camera the plot displays
a marker or a marker-cluster.
The result is almost the same as before. For overlapping markers we now
use OverlappingMarkerSpiderfier-Leaflet library.