[🐛 Bug Fix] Avoid computing diff of polygons to fix buildings import
The computed FOV of cameras are not donuts-shaped anymore, they are
circles. The diff between FOV is done either in raw SQL in the vector
tile view export (for list view) or in leaflet with outer-ring
substraction (for camera details display).

This fix a bug during import of nearby buildings in the installation
sequence.
This also simplify the backend code to speed up computation.